engine/requirements.txt
# NetScope AI Engine dependencies
# This process owns LLM, RAG, ChromaDB, and all agent tools.

# Redis bus
redis[hiredis]>=5.0

# LLM
openai>=1.30.1
httpx>=0.27.0

# Config
pydantic>=2.7.1
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# RAG stack
markitdown[pdf]>=0.1.5
chromadb>=1.5.0
sentence-transformers>=3.0.0
flashrank>=0.2.0
rank_bm25>=0.2.2
torch>=2.6.0

# MCP server
mcp>=1.0.0
